+++ commhandler/src/sgr_commhandler/driver/modbus/auxiliary_functions.py
