+++ commhandler/src/sgr_commhandler/driver/modbus/modbus_rtu_client_async.py
@@ -3,7 +3,7 @@ import asyncio
 from pymodbus.client import AsyncModbusSerialClient
 from pymodbus.constants import Endian
 
-from ..payload_decoder import (
+from payload_decoder import (
     PayloadBuilder,
     PayloadDecoder,
     RoundingScheme,
