+++ commhandler/src/sgr_commhandler/driver/modbus/payload_decoder.py
