+++ specification/setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name='SGrSpecificationPythontks4r',  # SGrSpecificationPythontks4r
+    name='SGrSpecificationPythontks4r',  # use SGrSpecificationPythontks4r until organization has been created, then change to sgr-specification
     version='0.3.0',
     packages=find_packages(where='src'),
     package_dir={'':'src'},
@@ -11,7 +11,8 @@ setup(
         'Issue Tracker': 'https://github.com/SmartGridready/SGrSpecifications/issues',
         'Source': 'https://github.com/SmartGridready/SGrSpecifications'
     },
-    license='BSD',
+    license='BSD-3',
+    license_files=['LICENSE.md'],
     author='Robin Schoch',
     author_email='',
     description='SGr Specifications',
